+++ tests/conftest.py
@@ -2,13 +2,7 @@ from pathlib import Path
 
 import pytest
 from fastapi.testclient import TestClient
-from qcelemental.models import (
-    AtomicInput,
-    AtomicResult,
-    Molecule,
-    OptimizationInput,
-    OptimizationResult,
-)
+from qcio import Molecule, ProgramInput
 
 from chemcloud_server.auth import bearer_auth
 from chemcloud_server.config import get_settings
@@ -55,7 +49,7 @@ def hydrogen():
         **{
             "symbols": ["H", "H"],
             "geometry": [0, 0, -0.65, 0.0, 0.0, 0.65],
-            "molecular_multiplicity": 1,
+            "multiplicity": 1,
             "connectivity": [[0, 1, 1]],
         }
     )
@@ -84,49 +78,13 @@ def water():
 
 
 @pytest.fixture
-def atomic_input(water):
-    return AtomicInput(
-        molecule=water, driver="energy", model={"method": "HF", "basis": "sto-3g"}
-    )
-
-
-@pytest.fixture
-def opt_input(atomic_input):
-    ai_dict = atomic_input.dict()
-    ai_dict.pop("molecule")
-    return OptimizationInput(
-        input_specification={
-            "driver": atomic_input.driver,
-            "model": atomic_input.model,
-        },
-        initial_molecule=atomic_input.molecule,
-    )
-
-
-@pytest.fixture
-def atomic_result(atomic_input):
-    return AtomicResult(
-        **atomic_input.dict(),
-        return_result=123.4,
-        success=True,
-        properties={},
-    )
-
-
-@pytest.fixture
-def opt_result(atomic_result, opt_input):
-    opt_dict = opt_input.dict()
-    opt_dict.pop("schema_name")
-    return OptimizationResult(
-        **opt_dict,
-        final_molecule=atomic_result.molecule,
-        trajectory=[atomic_result],
-        energies=[atomic_result.return_result],
-        success=True,
+def program_input(water):
+    return ProgramInput(
+        molecule=water, calctype="energy", model={"method": "HF", "basis": "sto-3g"}
     )
 
 
 @pytest.fixture(scope="session")
 def settings():
-    """QC Cloud application settings"""
+    """ChemCloud application settings"""
     return get_settings()